+++ src/io.hh
@@ -27,9 +27,9 @@
 
 
 /// \brief Écrit dans le fichier le texte compressé
-void write_file(FILE *const, const std::vector<std::vector<std::uint32_t>> &);
+void write_file(FILE *const, const std::vector<std::vector<std::uint16_t>> &);
 
 /// \brief Écrit un chunk dans le fichier de sortie
-void write_chunk(FILE *const, const std::vector<std::uint32_t> &);
+void write_chunk(FILE *const, const std::vector<std::uint16_t> &);
 
 #endif /* LZW_SRC_IO_H_ */
